Move query_params in commands folder to experimental_query_params.py
